Reorganized the file structure
